This refers to the crontab of linux
If you want to schedule some tasks that linux could execute by
themselves then crontab is a solution. However, I didn't manage to find
in the internet how  exactly to fix the code, so this is the code that
worked in the end.
There is also some example code that you can execute.